chore(rust): move the public tile model out of decoder into crate::tile
